requirements.txt
absl-py==0.10.0
appdirs==1.4.3
asgiref==3.2.10
astroid==2.4.2
astunparse==1.6.3
CacheControl==0.12.6
cachetools==4.1.1
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.0
distro==1.4.0
Django==3.1.1
djangorestframework==3.11.1
gast==0.3.3
google-auth==1.21.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
gunicorn==20.0.4
h5py==2.10.0
html5lib==1.0.1
idna==2.8
importlib-metadata==1.7.0
ipaddr==2.2.0
isort==5.5.3
itypes==1.2.0
Jinja2==2.11.2
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.4.3
lockfile==0.12.2
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==0.6.2
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.42
opt-einsum==3.3.0
packaging==20.3
pep517==0.8.2
Pillow==7.2.0
pkg-resources==0.0.0
progress==1.5
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.6.0
pyparsing==2.4.6
pytoml==0.1.21
pytz==2020.1
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
scipy==1.4.1
six==1.14.0
sqlparse==0.3.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
toml==0.10.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
